requirements.txt
appdirs==1.4.4
black==21.7b0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
ffmpeg-python==0.2.0
firebase-admin==5.0.1
future==0.18.2
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-firestore==2.2.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpcio==1.39.0
httplib2==0.19.1
idna==3.2
loguru==0.5.3
msgpack==1.0.2
mypy-extensions==0.4.3
numpy==1.21.1
opencv-python==4.5.3.56
packaging==21.0
pathspec==0.9.0
Pillow==8.3.1
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
regex==2021.7.6
requests==2.26.0
rsa==4.7.2
six==1.16.0
tomli==1.2.0
uritemplate==3.0.1
urllib3==1.26.6
youtube-dl==2021.6.6
